+++ examples/react/index.html
@@ -6,6 +6,6 @@
 <body>
 <!--script-->
 <div id="app"><!--element--></div>
-<script type="module" src="/entry/client.js"></script>
+<script type="module" src="/entry/client.jsx"></script>
 </body>
 </html>
