+++ examples/react/pages/hello.jsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from 'react'
-import { useSSEContext } from 'fastify-vite/react'
+import { useSSEContext } from 'fastify-vite-react/client'
 
-export default function Hello() {
-  const { context, setContext } = useSSEContext();
-  let [msg, setMsg] = useState(context?.$data?.message);
+export default function Hello () {
+  const { context, setContext } = useSSEContext()
+  const [msg, setMsg] = useState(context?.$data?.message)
 
   const refreshData = async () => {
     const response = await fetch(context.$dataPath())
@@ -13,10 +13,9 @@ export default function Hello() {
 
   useEffect(() => {
     if (!msg && !import.meta.env.SSR) {
-      refreshData();
+      refreshData()
     }
   })
 
   return <h1 onClick={refreshData}>{msg}</h1>
-
 }
